requirements.txt
# Personal Freedom - Requirements
# Install with: pip install -r requirements.txt

# For system tray functionality
pystray>=0.19.0
Pillow>=9.0.0
